backend/requirements.txt
flask
flask-cors
python-jose
requests
